修复音频头信息写入错误BUG

+++ minerva/src/main/java/com/lodz/android/minerva/bean/states/Recording.kt
@@ -5,4 +5,4 @@ package com.lodz.android.minerva.bean.states
  * @author zhouL
  * @date 2022/11/10
  */
-data class Recording(val db: Int, val data: ByteArray?) : RecordingStates()
+data class Recording(val db: Double, val data: ByteArray?) : RecordingStates()
